restructuring: make it more modular - move all chess related script to it's own folder (./chess)
