--- constants/chessConstants.js
@@ -1,3 +1,0 @@
-// All chess pieces variavble for comparison
-export let allWhitePieces = ['wp', 'wr', 'wn', 'wb', 'wq', 'wk'];
-export let allBlackPieces = ['bp', 'br', 'bn', 'bb', 'bq', 'bk'];
